refactor: Updated HOC error handler to use hooks
