+++ src/hoc/error-handler/index.js
@@ -1,38 +1,16 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import Modal from "../../components/ui/modal/";
+import useHttpErrorHandler from '../../hooks/http-error-handler'
 
 const withErrorHandler = (WrappedComponent, axios) => {
   return props => {
-    const [error, setError] = useState(null)
-
-    const requestInterceptor = axios.interceptors.request.use((req) => {
-      setError(null)
-      return req;
-    });
-
-    const responseInterceptor = axios.interceptors.response.use(
-      (res) => res,
-      (err) => {
-        setError(err)
-      }
-    );
-
-    useEffect(() => {
-      return () => {
-        axios.interceptors.request.eject(requestInterceptor);
-        axios.interceptors.response.eject(responseInterceptor);
-      }
-    }, [requestInterceptor, responseInterceptor])
-
-    const errorConfirmedHandler = () => {
-      setError(null)
-    };
+    const [error, clearError] = useHttpErrorHandler(axios)
 
     return (
       <React.Fragment>
         <Modal
           show={error}
-          modalClosed={errorConfirmedHandler}
+          modalClosed={clearError}
         >
           {error ? error.message : null}
         </Modal>
